requirements.txt
attrs==17.4.0
Automat==0.6.0
certifi==2018.1.18
chardet==3.0.4
constantly==15.1.0
diff-match-patch==20121119
Django==2.0.4
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.13
django-simple-captcha==0.5.6
et-xmlfile==1.0.1
future==0.16.0
googletrans==2.2.0
httplib2==0.9.2
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.1
Pillow==5.0.0
PyMySQL==0.8.0
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.4.3
